% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{asgs_key}
\alias{asgs_key}
\title{Generate a human-readable cache key from a specification for an ASGS request}
\usage{
asgs_key(
  geography,
  edition,
  reference_date,
  layer,
  where,
  filter_geom,
  predicate,
  ...
)
}
\value{
Returns a (somewhat) descriptive key.
}
\description{
Generate a human-readable cache key from a specification for an ASGS request
}
