% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{ced}
\alias{ced}
\title{Commonwealth Electoral Divisions (CEDs)}
\usage{
ced(
  edition = NULL,
  reference_date = NULL,
  layer = c("gen", "full", "point"),
  where = NULL,
  filter_geom = NULL,
  predicate = "intersects",
  cache = getOption("ozgs.cache"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Fetch ASGS geometries for Commonwealth Electoral Divisions (CEDs).
}
\details{
The following combinations of ASGS \code{edition} and \code{reference_date} are available:\tabular{llr}{
   geography \tab edition \tab reference_date \cr
   CED \tab 1 \tab 2011 \cr
   CED \tab 1 \tab 2012 \cr
   CED \tab 2 \tab 2016 \cr
   CED \tab 2 \tab 2017 \cr
   CED \tab 2 \tab 2018 \cr
   CED \tab 3 \tab 2021 \cr
}
}
