% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{iare}
\alias{iare}
\title{Indigenous Areas (IAREs)}
\usage{
iare(
  edition = NULL,
  reference_date = NULL,
  layer = c("gen", "full", "point"),
  where = NULL,
  filter_geom = NULL,
  predicate = "intersects",
  cache = getOption("ozgs.cache"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Fetch ASGS geometries for Indigenous Areas (IAREs).
}
\details{
The following combinations of ASGS \code{edition} and \code{reference_date} are available:\tabular{llr}{
   geography \tab edition \tab reference_date \cr
   IARE \tab 1 \tab 2011 \cr
   IARE \tab 2 \tab 2016 \cr
   IARE \tab 3 \tab 2021 \cr
}
}
