% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{lga}
\alias{lga}
\title{Local Government Areas (LGAs)}
\usage{
lga(
  edition = NULL,
  reference_date = NULL,
  layer = c("gen", "full", "point"),
  where = NULL,
  filter_geom = NULL,
  predicate = "intersects",
  cache = getOption("ozgs.cache"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Fetch ASGS geometries for Local Government Areas (LGAs).
}
\details{
The following combinations of ASGS \code{edition} and \code{reference_date} are available:\tabular{llr}{
   geography \tab edition \tab reference_date \cr
   LGA \tab 1 \tab 2011 \cr
   LGA \tab 1 \tab 2012 \cr
   LGA \tab 1 \tab 2013 \cr
   LGA \tab 1 \tab 2014 \cr
   LGA \tab 1 \tab 2015 \cr
   LGA \tab 2 \tab 2016 \cr
   LGA \tab 2 \tab 2017 \cr
   LGA \tab 2 \tab 2018 \cr
   LGA \tab 2 \tab 2019 \cr
   LGA \tab 2 \tab 2020 \cr
   LGA \tab 3 \tab 2021 \cr
   LGA \tab 3 \tab 2022 \cr
   LGA \tab 3 \tab 2023 \cr
}
}
