% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{sa4}
\alias{sa4}
\title{Statistical Areas Level 4 (SA4s)}
\usage{
sa4(
  edition = NULL,
  reference_date = NULL,
  layer = c("gen", "full", "point"),
  where = NULL,
  filter_geom = NULL,
  predicate = "intersects",
  cache = getOption("ozgs.cache"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Fetch ASGS geometries for Statistical Areas Level 4 (SA4s).
}
\details{
The following combinations of ASGS \code{edition} and \code{reference_date} are available:\tabular{llr}{
   geography \tab edition \tab reference_date \cr
   SA4 \tab 1 \tab 2011 \cr
   SA4 \tab 2 \tab 2016 \cr
   SA4 \tab 3 \tab 2021 \cr
}
}
