% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{sal}
\alias{sal}
\title{Suburbs and Localities (SALs)}
\usage{
sal(
  edition = NULL,
  reference_date = NULL,
  layer = c("gen", "full", "point"),
  where = NULL,
  filter_geom = NULL,
  predicate = "intersects",
  cache = getOption("ozgs.cache"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Fetch ASGS geometries for Suburbs and Localities (SALs).
}
\details{
The following combinations of ASGS \code{edition} and \code{reference_date} are available:\tabular{llr}{
   geography \tab edition \tab reference_date \cr
   SAL \tab 3 \tab 2021 \cr
}
}
