% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{sed}
\alias{sed}
\title{State Electoral Divisions (SEDs)}
\usage{
sed(
  edition = NULL,
  reference_date = NULL,
  layer = c("gen", "full", "point"),
  where = NULL,
  filter_geom = NULL,
  predicate = "intersects",
  cache = getOption("ozgs.cache"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Fetch ASGS geometries for State Electoral Divisions (SEDs).
}
\details{
The following combinations of ASGS \code{edition} and \code{reference_date} are available:\tabular{llr}{
   geography \tab edition \tab reference_date \cr
   SED \tab 1 \tab 2011 \cr
   SED \tab 2 \tab 2016 \cr
   SED \tab 2 \tab 2017 \cr
   SED \tab 2 \tab 2018 \cr
   SED \tab 2 \tab 2019 \cr
   SED \tab 2 \tab 2020 \cr
   SED \tab 3 \tab 2021 \cr
   SED \tab 3 \tab 2022 \cr
}
}
