% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{ste}
\alias{ste}
\title{State and Territory (STE)}
\usage{
ste(
  edition = NULL,
  reference_date = NULL,
  layer = c("gen", "full", "point"),
  where = NULL,
  filter_geom = NULL,
  predicate = "intersects",
  cache = getOption("ozgs.cache"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Fetch ASGS geometries for State and Territory (STE).
}
\details{
The following combinations of ASGS \code{edition} and \code{reference_date} are available:\tabular{llr}{
   geography \tab edition \tab reference_date \cr
   STE \tab 1 \tab 2011 \cr
   STE \tab 2 \tab 2016 \cr
}
}
