% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{tr}
\alias{tr}
\title{Tourism Regions (TRs)}
\usage{
tr(
  edition = NULL,
  reference_date = NULL,
  layer = c("gen", "full", "point"),
  where = NULL,
  filter_geom = NULL,
  predicate = "intersects",
  cache = getOption("ozgs.cache"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Fetch ASGS geometries for Tourism Regions (TRs).
}
\details{
The following combinations of ASGS \code{edition} and \code{reference_date} are available:\tabular{llr}{
   geography \tab edition \tab reference_date \cr
   TR \tab 1 \tab 2015 \cr
   TR \tab 2 \tab 2016 \cr
   TR \tab 3 \tab 2021 \cr
}
}
