% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{dzn}
\alias{dzn}
\title{Destination Zones (DZNs)}
\usage{
dzn(
  edition = NULL,
  reference_date = NULL,
  layer = c("gen", "full", "point"),
  where = NULL,
  filter_geom = NULL,
  predicate = c("intersects", "contains", "crosses", "overlaps", "touches", "within"),
  cache = getOption("ozgs.cache"),
  ...
)
}
\arguments{
\item{edition}{An ASGS edition: \code{1}, \code{2}, or \code{3}.}

\item{reference_date}{The geography's year of release. For most geographies,
\code{reference_year} is optional and specifying an \code{edition} will be
sufficient. However, for LGAs, CEDs and SEDs, the ASGS contains multiple
releases per edition. For these geographies, a \code{reference_date} must be
supplied to uniquely identify a release.}

\item{layer}{One of:
\itemize{
\item \code{"gen"}, the default. Fetches simplified geometries that have been generalised to 0.000025° or 2.5m.
\item \code{"full"}, full ASGS geometries. Identical to ASGS Shapefile and Geopackage downloads.
\item \code{"point"}, point geometries for each records in a geography.
}}

\item{where}{An optional SQL WHERE clause to filter the records returned by
the request.}

\item{filter_geom}{An optional \code{\link[sf:sfc]{sf::sfc}} or single \code{sf} geometry to filter
the records returned by the request.}

\item{predicate}{An optional spatial predicate to specify the relation
between \code{filter_geom} and \code{geography}. One of \code{"intersects"}, \code{"contains"},
\code{"crosses"}, \code{"overlaps"}, \code{"touches"}, and \code{"within"}.}

\item{cache}{A \code{cachem}-compatible cache. If not otherwise specified,
defaults to a memory cache. To persistently store downloaded ASGS
geometries and data, supply an object created by \code{\link[cachem:cache_disk]{cachem::cache_disk()}}.}

\item{...}{Additional arguments passed to \code{\link[arcgislayers:arc_read]{arcgislayers::arc_read()}}.}
}
\description{
Fetch ASGS geometries for the \strong{Destination Zones (DZNs)} geography.
}
\details{
Destination Zones are co-designed with state and territory transport authorities. DZNs are designed from either whole or aggregations of Mesh Blocks (MB) and they align to Statistical Area Level 2 boundaries. They have no relationship to Statistical Area Level 1 boundaries. Destination Zones were incorporated as part of the ASGS for the first time in 2021. Further information about these regions can be found in the following publication: Australian Statistical Geography Standard (ASGS) Edition 3 https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026
}
\section{Availability}{

The following combinations of ASGS \code{edition} and \code{reference_date} are available:\tabular{llr}{
   \strong{\code{geography}} \tab \strong{\code{edition}} \tab \strong{\code{reference_date}} \cr
   DZN \tab 2 \tab 2016 \cr
   DZN \tab 3 \tab 2021 \cr
}
}

