% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geography.R
\name{get_geography}
\alias{get_geography}
\title{Fetch a geography from the ASGS web service}
\usage{
get_geography(
  geography,
  edition = NULL,
  reference_date = NULL,
  layer = c("gen", "full", "point"),
  where = NULL,
  filter_geom = NULL,
  predicate = "intersects",
  cache = getOption("ozgs.cache"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Fetch a geography from the ASGS web service
}
