% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{ra}
\alias{ra}
\title{Remoteness Areas (RAs)}
\usage{
ra(
  identifier = NULL,
  edition = NULL,
  reference_date = NULL,
  layer = c("gen", "full", "point"),
  where = NULL,
  filter_geom = NULL,
  predicate = c("intersects", "contains", "crosses", "overlaps", "touches", "within"),
  cache = getOption("ozgs.cache"),
  ...
)
}
\arguments{
\item{identifier}{An optional character vector of named features to download
from the specified geography. For most ASGS geographies, these are
human-friendly names like e.g. "Tasmania" (a feature within the STE
geography), "Sydney" (a feature within the LGA geography) or "2150" (a
feature within the POA geography). If specified, \code{identifier} takes
precedence over \code{where}. If neither \code{identifier} nor \code{where} are specified,
all of the features in the specified geography will be downloaded.}

\item{edition}{An ASGS edition: \code{1}, \code{2}, or \code{3}.}

\item{reference_date}{The geography's year of release. For most geographies,
\code{reference_year} is optional and specifying an \code{edition} will be
sufficient. However, for LGAs, CEDs and SEDs, the ASGS contains multiple
releases per edition. For these geographies, a \code{reference_date} must be
supplied to uniquely identify a release.}

\item{layer}{One of:
\itemize{
\item \code{"gen"}, the default. Fetches simplified geometries that have been
generalised to 0.000025° or 2.5m.
\item \code{"full"}, full ASGS geometries. Identical to ASGS Shapefile and Geopackage
downloads.
\item \code{"point"}, point geometries for each records in a geography.
}}

\item{where}{An optional SQL WHERE clause to filter the features returned by
the request. Ignored if \code{identifier} is specified.}

\item{filter_geom}{An optional \code{\link[sf:sfc]{sf::sfc}} or single \code{sf} geometry to filter
the records returned by the request.}

\item{predicate}{An optional spatial predicate to specify the relation
between \code{filter_geom} and \code{geography}. One of \code{"intersects"}, \code{"contains"},
\code{"crosses"}, \code{"overlaps"}, \code{"touches"}, and \code{"within"}.}

\item{cache}{A \code{cachem}-compatible cache. If not otherwise specified,
defaults to a memory cache. To persistently store downloaded ASGS
geometries and data, supply an object created by \code{\link[cachem:cache_disk]{cachem::cache_disk()}}.}

\item{...}{Additional arguments passed to \code{\link[arcgislayers:arc_read]{arcgislayers::arc_read()}}.}
}
\description{
Fetch ASGS geometries for the \strong{Remoteness Areas (RAs)} geography.
}
\details{
The Remoteness Areas divide Australia into 5 classes of remoteness on the basis of relative access to services for the purpose of releasing and analysing statistical data. Access to services is measured using the Accessibility and Remoteness Index of Australia (ARIA+), provided by the Hugo Centre for Migration and Population Research at the University of Adelaide. Remoteness Areas are aggregates of Statistical Areas Level 1 (SA1s) that are grouped together based on their average ARIA+ score. These areas are redefined every five years and data from the Census of Population and Housing as well as other survey data are released on the Remoteness Areas.
}
\section{Availability}{

The following combinations of ASGS \code{edition} and \code{reference_date} are available:\tabular{llr}{
   \strong{\code{geography}} \tab \strong{\code{edition}} \tab \strong{\code{reference_date}} \cr
   RA \tab 1 \tab 2011 \cr
   RA \tab 2 \tab 2016 \cr
   RA \tab 3 \tab 2021 \cr
}
}

